% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_fisheye.R
\name{remove_fisheye}
\alias{remove_fisheye}
\title{Remove fisheye}
\usage{
remove_fisheye(file = NULL, dir = NULL, output = NULL, remove_audio = FALSE)
}
\arguments{
\item{file}{string. Name of file in `dir` or current working directory.}

\item{dir}{string. Directory location of file you want to trim. If left NULL
uses current working directory.}

\item{output}{string. Name of new file, without filetype extension. If left
NULL output is named same as the original file appended with `_no_fisheye`.}

\item{remove_audio}{logical. If TRUE audio is removed from the copy.}
}
\description{
Remove fisheye effect from a GoPro Hero Black 5 or 6
}
\examples{
# Commented out because we don't want package checks to actually run this
#
# remove_fisheye(file = "path/to/file",
#                dir = "path/to/dir",
#                output = "output",
#                remove_audio = FALSE)
}
