% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_vid.R
\name{trim_vid}
\alias{trim_vid}
\title{trim_vid}
\usage{
trim_vid(file = NULL, dir = NULL, output = NULL, from = NULL,
  to = NULL)
}
\arguments{
\item{file}{string. Name of file in `dir` or current working directory.}

\item{dir}{string. Directory location of file you want to trim. If left NULL
uses current working directory.}

\item{output}{string. Name of new file, without filetype extension. If left
NULL output is named same as the original file appended with `_trim`.}

\item{from}{string. Time into trim from in hh:mm:ss format.}

\item{to}{string Time to trim to in hh:mm:ss format.}
}
\description{
Trims a movie file to the given `from` and `to` timepoints.
}
\details{
Either have the file in the current working directory, or specify the
location with `dir`, in which case the working directory is unchanged.

Can rename the output file using `output`, otherwise it will be named the
same as the original file appended with `_trim`. It will overwrite any
previous trimmed file with the same name.

No format conversion or rencoding is done. New file is same format as old.
}
\examples{
# Commented out because we don't want package checks to actually run this
#
# trim_vid(dir = "path/to/directory",
#          file = "input_file.mp4",
#          output = "output",
#          from = "00:02:45",
#          to = "00:09:15")
}
